requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==6.0.0
idna==3.6
mysql-connector-python==8.3.0
pyTelegramBotAPI==4.16.1
python-telegram-bot==13.7
pytube==15.0.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
setuptools==69.1.0
six==1.16.0
telebot==0.0.5
telegram==0.0.1
tornado==6.4
tzlocal==5.2
urllib3==1.26.18
